TRACK-466 Simplify SearchTable declarations
In preparation for adding a bunch of new search tables, make it less
cumbersome to declare them by providing more default values and adding
a helper base class for tables with per-facility permissions.

Move SearchTables out of the seedbank package since it is going to
have tables other than seed-bank-related ones.
