+++ src/main/kotlin/com/terraformation/backend/search/namespace/CollectorsNamespace.kt
@@ -2,9 +2,9 @@ package com.terraformation.backend.search.namespace
 
 import com.terraformation.backend.db.tables.references.COLLECTORS
 import com.terraformation.backend.search.SearchFieldNamespace
+import com.terraformation.backend.search.SearchTables
 import com.terraformation.backend.search.SublistField
 import com.terraformation.backend.search.field.SearchField
-import com.terraformation.backend.seedbank.search.SearchTables
 
 class CollectorsNamespace(searchTables: SearchTables) : SearchFieldNamespace() {
   override val sublists: List<SublistField> = emptyList()
