+++ src/main/kotlin/com/terraformation/backend/search/namespace/GerminationTestsNamespace.kt
@@ -4,9 +4,9 @@ import com.terraformation.backend.db.tables.references.ACCESSIONS
 import com.terraformation.backend.db.tables.references.GERMINATIONS
 import com.terraformation.backend.db.tables.references.GERMINATION_TESTS
 import com.terraformation.backend.search.SearchFieldNamespace
+import com.terraformation.backend.search.SearchTables
 import com.terraformation.backend.search.SublistField
 import com.terraformation.backend.search.field.SearchField
-import com.terraformation.backend.seedbank.search.SearchTables
 
 class GerminationTestsNamespace(
     searchTables: SearchTables,
